AI Agent/requirements.txt
openai
torch
transformers
elasticsearch
pillow
streamlit
